All-in-one function to call the major steps of epitopefindr.

All-in-one function to call the major steps of epitopefindr.

epfind(
  data = NULL,
  output.dir = NULL,
  e.thresh = 0.01,
  g.method = "any",
  aln.size = 7,
  min.groupsize = 2,
  min.consensus.pos = 1,
  consensus.type = "Biostrings",
  consensus.thresh = c(100, 49),
  peptide.nchar = 50,
  msa.width = "dynamic",
  verbose = TRUE,
  pdflatex = TRUE,
  pdftk = FALSE,
  pdfuniter = TRUE,
  make.png = FALSE,
  name.msa = "msa.pdf",
  name.alignments = "finalAlignments.csv",
  name.epitopekey = "epitopeKey.csv",
  name.epitopesum = "epitopeSummary.csv",
  use.doParallel = FALSE,
  delete.intermediates = FALSE
)

epFind2(
  data = NULL,
  output.dir = NULL,
  e.thresh = 0.01,
  g.method = "any",
  aln.size = 7,
  min.groupsize = 2,
  min.consensus.pos = 1,
  consensus.type = "Biostrings",
  consensus.thresh = c(100, 49),
  peptide.nchar = 50,
  msa.width = "dynamic",
  verbose = TRUE,
  pdflatex = TRUE,
  pdftk = FALSE,
  pdfuniter = TRUE,
  make.png = FALSE,
  name.msa = "msa.pdf",
  name.alignments = "finalAlignments.csv",
  name.epitopekey = "epitopeKey.csv",
  name.epitopesum = "epitopeSummary.csv",
  use.doParallel = FALSE,
  delete.intermediates = FALSE
)

Arguments

data

Biostrings::AAStringset input sequences to search for epitopes, or path to corresponding .fasta file.

output.dir

Directory to which output files should be written.

e.thresh

Maximum e-value to consider from BLASTp alignments of 'data'.

g.method

Grouping method of alignments. Either 'any' or 'all'. See ?indexGroups

aln.size

Minimum length of alignment to consider from BLASTp alignments of 'data'.

min.groupsize

Minimum number of peptides per group to require in order to print a group.

min.consensus.pos

Minimum number of amino acid consensus positions required in order to print a group.

consensus.type

"upperlower" or "Biostrings" type of msa consensus sequence to output

consensus.thresh

Two decreasing numeric values of upper and lower thresholds for sequence consensus.

peptide.nchar

Maximum of character from peptide name to use in msa output. Default 50. Starts from left.

msa.width

Controls whether or not MSA images have fixed or dynamic width. By default, msa.width is set to "dynamic", which causes the document dimentions of the resultant image to be calculated based on the lentht of the peptide name and the number of amino acids in the sequence alignment. If msa.width is instead set to a numeric, then an MSA will be printed with a fixed with that number of inches. With 50-character peptide.nchar and a maximum expected sequence alignment of 45 positions, an msa.width of 12 is more than sufficient.

verbose

Logical to print progress updates.

pdflatex

Logical whether or not to produce PDF LaTeX figures using pdflatex

pdftk

Logical whether or not to merge msa pdfs using staplr and pdftk

pdfuniter

Logical whether or not to merge msa pdfs using staplr and pdfuniter

make.png

Locial whether or not to convert PDF output to PNG.

name.msa

Filename for output merged pdf of msa logos.

name.alignments

Filename for output spreadsheet of peptide alignments.

name.epitopekey

Filename for output spreadhseet of epitopes per peptide.

name.epitopesum

Filename for output summary sheet of epitopes.

use.doParallel

Logical whether or not to use doParallel parallelization.

delete.intermediates

Logical whether or not to delete the intermediate_files folder at the end